Scaled stuff

+++ PA11/include/object.h
@@ -20,7 +20,7 @@ class Object {
 	public:
 		//The characteristics of a particular planet that can be changed at any time
 		struct Context {
-            bool isDynamic;
+			bool isDynamic;
 			bool isBall = false;
 			bool isAlt = false;
 			
@@ -29,7 +29,7 @@ class Object {
 			std::string name = "Object";
 			std::string vertexShader;
 			std::string fragmentShader;
-
+			
 			Model* model = nullptr;
 			Texture* texture = nullptr;
 			Texture* altTexture = nullptr;
@@ -38,10 +38,10 @@ class Object {
 			Shader* shader;
 			Shader* altShader;
 			
-            int xLoc = 0;
-            int yLoc = 0;
-            int zLoc = 0;
-
+			float xLoc = 0;
+			float yLoc = 0;
+			float zLoc = 0;
+			
 			int shape = 0;
 			float mass = 0;
 			
@@ -49,28 +49,32 @@ class Object {
 			glm::vec3 tempScale = glm::vec3(1.0f, 1.0f, 1.0f);
 			
 			int id = 0;
-
-            bool hasPhysics = true;
+			
+			bool hasPhysics = true;
 			bool isBounceType = false;
-            btRigidBody * physicsBody;
+			btRigidBody* physicsBody;
 			int rigidBodyIndex = 0;
-
+			
 			int expansionTimer = 0;
 		};
 		
-		Object(const Context &ctx);
+		Object(const Context& ctx);
 		
 		~Object();
 		
 		//Initialises the planet's model and textures for OpenGL
-		void Init_GL(std::unordered_map<std::string, std::string> const * dictionary = nullptr);
+		void Init_GL(std::unordered_map<std::string, std::string> const* dictionary = nullptr);
+		
 		//Updates the physics for the planet
 		void Update(float dt);
+		
 		//Renders the planet on the screen
 		void Render(bool withShadows, std::vector<glm::mat4> spotlightMatrices) const;
+		
 		void RenderID(Shader* shader) const;
+		
 		void RenderShadow(Shader* shader, const glm::mat4& lightMatrix) const;
-
+		
 		//Returns the current model matrix of this planet
 		const glm::mat4& GetModel() const;
 		
@@ -81,10 +85,10 @@ class Object {
 		
 		//Where in the world the planet is
 		const glm::vec3& position;
-
+		
 		static glm::mat4* viewMatrix;
 		static glm::mat4* projectionMatrix;
-
+		
 		static Menu* menu;
 	
 	protected:
